requirements.txt
PyYAML
numpy
pygame
